--- tests/Dockerfile.mamba
@@ -1,18 +1,0 @@
-FROM continuumio/miniconda3
-
-RUN conda install mamba -c conda-forge -y
-
-WORKDIR /app
-
-ADD . /app
-
-# Python
-RUN pip install cookiecutter pre-commit pytest pytest-order
-
-# Latest stable terraform
-RUN apt-get update && apt-get install -y gnupg software-properties-common wget unzip \
-    && wget -O- https://apt.releases.hashicorp.com/gpg | gpg --dearmor | tee /usr/share/keyrings/hashicorp-archive-keyring.gpg \
-    && echo "deb [signed-by=/usr/share/keyrings/hashicorp-archive-keyring.gpg] https://apt.releases.hashicorp.com $(lsb_release -cs) main" | tee -a /etc/apt/sources.list.d/hashicorp.list \
-    && apt-get update && apt-get install -y terraform
-
-CMD [ "pytest", "-m", "python_env"]
