+++ tests/__init__.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+TEST_OUTPUT_DIR = Path("test_output")
