--- tests/docker-compose.yaml
@@ -1,19 +1,0 @@
-version: "3"
-services:
-  test-default:
-    platform: linux/x86_64
-    build:
-      context: .
-      dockerfile: Dockerfile
-    volumes:
-      - ../:/app
-    working_dir: /app
-
-  test-python_envs:
-    platform: linux/x86_64
-    build:
-      context: .
-      dockerfile: Dockerfile.mamba
-    volumes:
-      - ../:/app
-    working_dir: /app
